Attempt to automatically configure release notes
This commit is an attempt to tackle #7068 by configuring the release
notes in Github Releases with the handwritten release notes from
`RELEASES.md`. The basic idea here is to split the markdown file on
`-----` delimiters and then find the one which matches the version being
released. Once one is found the `body` field of the API call to create
the release is configured.

This is unfortunately quite difficult to test, so I haven't actually
tested it. I'm hoping this won't at least halt the release so we can see
how it goes for the next one and fix any issues that arise.